ai_service/requirements.txt
flask
pymongo
python-dotenv
numpy
torch
